# Replace `sha256` with `blake3` hash in image-optimizer