Here is benchmark suite for the `image-optimizer` scenario: https://github.com/Brooooooklyn/blake-hash/blob/main/bench-next.mjs

And the result shows `Blake3` hasher is much more faster than `sha256`:

### Calculate hash for `[3, 'http://abc.xyz/logo.webp', 256, 0.9, 'image/webp']`

```text
Running "digest hash into url-safe-base64" suite...
Progress: 100%

  blake3:
    1 512 961 ops/s, ±1.87%   | fastest

  blake2b:
    276 321 ops/s, ±2.97%     | slowest, 81.74% slower

  blake2s:
    290 832 ops/s, ±3.23%     | 80.78% slower

  sha256:
    296 197 ops/s, ±4.83%     | 80.42% slower

Finished 4 cases!
  Fastest: blake3
  Slowest: blake2b
```

### Calculate hash for a [png file](https://github.com/Brooooooklyn/blake-hash/blob/main/anime-girl.png)

```text
Running "digest big file" suite...
Progress: 100%

  blake3:
    6 729 ops/s, ±1.14%   | fastest

  blake2b:
    1 240 ops/s, ±0.61%   | 81.57% slower

  blake2bp:
    2 251 ops/s, ±2.40%   | 66.55% slower

  blake2s:
    722 ops/s, ±1.27%     | 89.27% slower

  blake2sp:
    2 417 ops/s, ±1.61%   | 64.08% slower

  sha256:
    468 ops/s, ±1.43%     | slowest, 93.05% slower

Finished 6 cases!
  Fastest: blake3
  Slowest: sha256
```