@Brooooooklyn 2021-11-10
> 
> 
> Did you compare the performance to any other built-in hashes?
> 
> I would prefer to avoid adding another dependency, especially if it means we're going to invalidate the cache for all Next.js deployments.

From this article: https://medium.com/asecuritysite-when-bob-met-alice/blake3-3716708235ac, `Blake3` is significantly faster than the other security hash algorithm.