@styfle 2021-11-10
Whats the digest size? Is this still 256 bit?