@Brooooooklyn 2021-11-10
> 
> 
> Whats the digest size? Is this still 256 bit?

Yes, `blake3` is 256 bit https://en.wikipedia.org/wiki/BLAKE_(hash_function)#BLAKE3