@creationix 2021-11-10
Assuming I understand the use case correctly here, I suggest we also compare non-cryptographic hashes.  In particular, I recommend adding the native and wasm versions and seeing how they compare.  I'm sure the second one is faster, but the first may be plenty fast and it's less headache being webassembly instead of native.

- https://www.npmjs.com/package/xxhash-wasm
- https://www.npmjs.com/package/xxhash-addon

As far as the chances of collisions I understand there are two keyspaces in which we want to avoid collisions:
- Each application has a local .next folder and we want all unique images to have unique hashes or the wrong image will be shown sometimes.  This should stay well below 1,000,000 unique values and is scoped to a single security domain (attacks would be against self)
- Browsers use the etag to know when to reuse a value in their cache.  This is scoped even more narrowly since it's per url.

Assuming a high quality non-cryptographic 64 bit hash the most likely collision is in the .next folder and it's about a 1-in-50,000,000 chance of there being a single collision among all of 1,000,000 unique images in that single application.  I consider this plenty safe considering the worse case is that 1 in 50 million mega sites may have one image rendered wrong.  A more likely scenario is a site with 2,000 unique images which has a 1-in-10,000,000,000,000 change of a single collision.

The etag is a non issue since there are unlikely to ever be more than two images for the same url in the browser cache so the chances of that collision are more like 1-in-18,000,000,000,000,000,000