@creationix 2021-11-10
If we really want to stick to built-in hashes in openssl and really care about performance, then maybe consider blake2 as it's pretty fast compared to sha256 and has a variable hash size as a supported feature so you could chose 64 bits and still be safe enough from collisions.