@Brooooooklyn 2021-11-10
I will add [xxhash](https://github.com/napi-rs/node-rs/blob/main/packages/xxhash/README.md) to benchmark tomorrow