@timneutkens 2021-11-11
@Brooooooklyn just as a FYI webpack has xxhash64 using wasm in the latest version as an opt-in option and is enabled in Next.js