@Brooooooklyn 2021-11-11
### Benchmark result with `xxhash` and `node:blake2b512`

<details>
<summary>Calculate hash for a [3, 'http://abc.xyz/logo.webp', 256, 0.9, 'image/webp']
</summary>
<p>

```text
Running "digest hash into url-safe-base64" suite...
  blake3:
    1 730 591 ops/s, ±1.64%   | fastest

  blake2b:
    288 755 ops/s, ±2.21%     | slowest, 83.31% slower

  blake2s:
    292 547 ops/s, ±2.89%     | 83.1% slower

  node:sha256:
    324 604 ops/s, ±4.30%     | 81.24% slower

  node:blake2b512:
    301 157 ops/s, ±2.49%     | 82.6% slower

  xxh3:
    569 721 ops/s, ±1.44%     | 67.08% slower

  xxh64:
    544 897 ops/s, ±1.84%     | 68.51% slower

Finished 7 cases!
  Fastest: blake3
  Slowest: blake2b
```

</p>
</details>  


<details>
<summary>Calculate hash for 888kb PNG file</summary>
<p>

```text
Running "digest big file" suite...
  blake3:
    6 814 ops/s, ±1.13%    | 60.93% slower

  blake2b:
    1 261 ops/s, ±0.34%    | 92.77% slower

  blake2bp:
    2 329 ops/s, ±1.32%    | 86.65% slower

  blake2s:
    745 ops/s, ±0.44%      | 95.73% slower

  blake2sp:
    2 477 ops/s, ±1.08%    | 85.8% slower

  node:sha256:
    470 ops/s, ±1.12%      | slowest, 97.31% slower

  node:blake2b512:
    826 ops/s, ±1.64%      | 95.26% slower

  xxh64:
    13 583 ops/s, ±1.97%   | 22.12% slower

  xxh3:
    17 441 ops/s, ±1.30%   | fastest

Finished 9 cases!
  Fastest: xxh3
  Slowest: node:sha256
```

</p>
</details> 


`xxh3` is ***almost 3x faster*** than `blake3` in the **Big Image file**  scenario, and more than ***3x slower*** than `blake3`.

The `node:blake2b512` is almost the same performance with `node:sha256` in the small input scenario and ***2x faster*** in the big file input scenario.

> Assuming a high quality non-cryptographic 64 bit hash, the most likely collision is in the .next folder and it's about a 1-in-50,000,000 chance of there being a single collision among all of 1,000,000 unique images in that single application. I consider this plenty safe considering the worse case is that 1 in 50 million mega sites may have one image rendered wrong. 

I agree with it. 

If non-cryptographic 64 bit hash is a choice, I'll recommend https://github.com/tkaitchuck/ahash which is even faster than `xxhash`, then I'll develop Node.js binding for it.

If the security hash algorithm is still chosen, `Blake3` is the best at this moment.