@Brooooooklyn 2021-11-11
> @Brooooooklyn just as a FYI webpack has xxhash64 using wasm in the latest version as an opt-in option and is enabled in Next.js

I've been aware of this. The benchmark in [xxhash](https://github.com/napi-rs/node-rs/blob/main/packages/xxhash/README.md) is compare the native version to `webpack` verison: https://github.com/napi-rs/node-rs/blob/main/packages/xxhash/benchmark/xxhash.js#L6. From the result, `wasm` is almost  `30%` slower  than the native version.