@styfle 2021-11-11
I like using the same algorithm as webpack so I would vote for xxhash. Ideally we could stick to one algorithm and one package for all file hashing, not just images here.

Do you agree @timneutkens @creationix ? Also curious if we can replace this in other places.