@creationix 2021-11-11
I also like using xxhash since we use it other places as well already.